refactor: lógica de criação de usuário passada para o contexto de usuário
