+++ package.json
@@ -13,6 +13,7 @@
     "@types/node": "20.2.1",
     "@types/react": "18.2.6",
     "@types/react-dom": "18.2.4",
+    "@types/uuid": "^9.0.1",
     "autoprefixer": "10.4.14",
     "next": "13.4.3",
     "postcss": "8.4.23",
